aider: Add italicized list of timestamps at the end of each paragraph in the template. Each timestamp is a clickable link to set the audio player's offset.
# Aider chat conversation:

USER: Rather than having each paragraph of a chapter clickable to the timestamp, I want the end of the paragraph to include to show a italisized list of times associated with the paragraph (sourcIds) link that you can click on to jump to that portion of the audio using the data-start attribute.

ASSISTANT: ```diff
--- summarizer/template.html
+++ summarizer/template.html
@@ ... @@
-            <div class="cursor-pointer hover:bg-slate-50">${marked.parse(summary.markdown)}</div>
+            <div>${marked.parse(summary.markdown)} <i>${summary.sourceIds.map(id => `<a href="#" onclick="setAudioPlayerOffset('${TRANSCRIPT[id].start}')" class="timestamp-link">${TRANSCRIPT[id].start}</a>`).join(', ')}</i></div>
             <div class="transcripts"></div>
           `;
             chapterDiv.appendChild(summaryDiv);
           });
         });
       }
```

This diff will modify the template to include a list of italicized timestamps at the end of each paragraph. Each timestamp is a clickable link that will set the audio player's offset to the corresponding time.